Bug fix: Overlapping strands
Function RemoveConflict search
on the list of base pairs which
are not really added to the helix
structure, so, they can be unplugged
and connected after the folding.

This is a temporary solver. If the ovelapping
is such that the structure goes over another,
it will still link them in a non conventional way.
Here, it should be splitted.

Anyway, unintentional overlapping is avoided,
in which with previous version was not possible,
and a big intentional gap was required.